[bashrc] Make machine-specific setup hook-based
This splits up the allowable machine-specific code in the bashrc file
into three places, allowing finer-grained control over when changes are
introduced.  This allows (for example) some environment variables to be
set before any version-based aliases are defined, so if (as on Mac) the
GNU versions of coreutils aren't the default until after modifications
to the PATH, then aliases need to come after there.

Unifying all the possible requirements for different machines into one
location for the additional sourcing won't ever be possible, so the
hook-based setup is nicer.

This also allows (in the future) for the creation of toggle-switches and
settings for the bashrc file, defined by machine-specific files.