[vim] Extend python syntax highlighting
This adds a few extra niceties to python syntax highlighting, such as
lexical function highlighting (e.g. the use of "my_own_function()" will
now match in the code) and class highlighting.

References to "self" are now made to look like comments to remove the
clutter of initialisation functions.

The docstring folding function is moved in to the python after file, for
separation of concerns.