[vim] Add custom TeX syntax highlighting
I got sick of the default Vim syntax highlighting for LaTeX being too
slow, so I wrote my own.  This is much more bare-bones than the previous
one, but it has everything that I actually want (or that I could think
of while writing it).

It's now much much faster - scrolling speed is far far smoother, and
because there are fewer colours in use, it's much easier to understand
too.